Extracted activation of e2ee to a hook
